% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeinfCode.R
\name{compute_B}
\alias{compute_B}
\title{Compute B Matrix for Basis Functions}
\usage{
compute_B(z, d, pscoreZ, bases)
}
\arguments{
\item{z}{The covariate value (unused in the function but included for consistency).}

\item{d}{Binary treatment indicator (0 or 1).}

\item{pscoreZ}{The propensity score for treatment assignment given \code{Z}.}

\item{bases}{A list of basis functions, where:
\itemize{
\item \code{bases[[1]]$ib}: Basis functions for the control group (\code{d = 0}).
\item \code{bases[[2]]$ib}: Basis functions for the treated group (\code{d = 1}).
}}
}
\value{
A numeric vector representing the computed B matrix for the given inputs.
}
\description{
This function computes the B matrix based on the treatment indicator \code{d},
the propensity score \code{pscoreZ}, and a set of basis functions.
}
