% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeinfCode.R
\name{compute_C0_C1}
\alias{compute_C0_C1}
\title{Compute C0 and C1 Matrices}
\usage{
compute_C0_C1(data, bases)
}
\arguments{
\item{data}{A data frame containing outcome \code{y}, treatment \code{d}, and covariate \code{z}.}

\item{bases}{A list of basis functions.}
}
\value{
A list with elements:
\itemize{
\item \code{C0}: Vector of estimated C0 values.
\item \code{C1}: Matrix of estimated C1 values.
}
}
\description{
Computes the C0 and C1 matrices as defined in Torgovitsky (2020).
}
