% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeinfCode.R
\name{compute_gamma_lambda_df}
\alias{compute_gamma_lambda_df}
\title{Generate Gamma-Lambda Data Frame}
\usage{
compute_gamma_lambda_df(kappas, sample.size, MB = 1)
}
\arguments{
\item{kappas}{Vector of kappa values.}

\item{sample.size}{Integer; sample size.}

\item{MB}{Integer; 1 for endpoint method, 2 for equally spaced gamma values.}
}
\value{
A data frame with columns \code{gamma}, \code{lambda}, and \code{kappa}.
}
\description{
Generates a data frame of gamma and lambda values for use in optimization.
}
