% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeinfCode.R
\name{compute_QLC_nQstar_n}
\alias{compute_QLC_nQstar_n}
\title{Compute Quadratic, Linear, and Constant Terms}
\usage{
compute_QLC_nQstar_n(n, C0, C0.star, C1, C1.star, gamma_n, lambda_n)
}
\arguments{
\item{n}{Sample size.}

\item{C0}{Vector of C0 values (using sample data).}

\item{C0.star}{Vector of C0 values (using bootstrap data).}

\item{C1}{Matrix of C1 values (using sample data).}

\item{C1.star}{Matrix of C1 values (using bootstrap data).}

\item{gamma_n}{Scalar gamma value.}

\item{lambda_n}{Scalar lambda value.}
}
\value{
A list containing:
\itemize{
\item \code{quadratic}: Quadratic term matrix.
\item \code{linear}: Linear term vector.
\item \code{constant}: Constant term scalar.
}
}
\description{
Computes the QLC terms used in quadratic optimization for hypothesis testing.
}
