% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeinfCode.R
\name{compute_tau}
\alias{compute_tau}
\title{Compute Tau Matrix}
\usage{
compute_tau(data, dgp, bases, target.parameter, late.lb, late.ub)
}
\arguments{
\item{data}{A data frame containing the observed sample data. Should include columns:
\itemize{
\item \code{y}: Outcome variable.
\item \code{d}: Treatment indicator (0 or 1).
\item \code{z}: Covariate.
}
Set to NULL if using \code{dgp}.}

\item{dgp}{A data-generating process object containing the probability distribution of \code{Z}.
Set to NULL if using \code{data}.}

\item{bases}{A list of basis functions, where:
\itemize{
\item \code{bases[[1]]$ib}: Basis functions for the control group (\code{d = 0}).
\item \code{bases[[2]]$ib}: Basis functions for the treated group (\code{d = 1}).
}}

\item{target.parameter}{The treatment effect parameter to estimate (e.g., ATT, ATE, LATE).}

\item{late.lb}{Lower bound for LATE estimation (required when target.parameter = "LATE")}

\item{late.ub}{Upper bound for LATE estimation (required when target.parameter = "LATE")}
}
\value{
A numeric matrix representing the computed tau values.
}
\description{
This function calculates the tau matrix as discussed in Torgovitsky (2023).
It determines weighted averages of basis functions over the sample or a given DGP.
}
