% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeinfCode.R
\name{compute_test_stat}
\alias{compute_test_stat}
\title{Compute Test Statistics in Zhu (2020)}
\usage{
compute_test_stat(
  data,
  bases,
  beta.null,
  C0,
  C1,
  tau,
  return.gurobi = FALSE,
  gurobi.tries
)
}
\arguments{
\item{data}{A data frame or tibble that must contain the following columns:
\describe{
\item{\code{y}}{Binary outcome variable (0/1).}
\item{\code{d}}{Binary treatment indicator (0/1).}
\item{\code{z}}{Binary instrument indicator (0/1).}
}}

\item{bases}{A list of basis functions.}

\item{beta.null}{A numeric vector of target parameter values for which the test
statistic is solved.}

\item{C0}{Vector of C0 values (using sample data).}

\item{C1}{Matrix of C1 values (using sample data).}

\item{tau}{A numeric matrix representing the computed tau values.}

\item{return.gurobi}{Logical; if TRUE, returns raw Gurobi output.}

\item{gurobi.tries}{Integer; number of attempts to optimize using Gurobi.}
}
\value{
A tibble where each row corresponds to one \code{beta.null} from the
input vector \code{beta.nulls}. The returned columns are:
\itemize{
\item \code{t.value}: The target parameter value used in the constraint.
\item \code{test.stat}: The minimized QP objective value.
If the problem is infeasible, this is \code{NA}.
\item \code{gurobi.result}: A list-column containing the raw Gurobi output
for that \code{beta.null}.
}
}
\description{
The function computes the test statistic for each value in \code{beta.nulls}.
Gurobi is used as the QP solver. Change beta.nulls to beta.null?
}
