% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeinfCode.R
\name{expectBBPrime}
\alias{expectBBPrime}
\title{Expected B B' Matrix Calculation}
\usage{
expectBBPrime(dgp, bases)
}
\arguments{
\item{dgp}{A list describing the data-generating process.}

\item{bases}{A list of basis functions.}
}
\value{
A matrix representing the expected outer product of B matrices.
}
\description{
Computes the expected outer product of B matrices given a DGP and basis functions.
}
