% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeinfCode.R
\name{expectBY}
\alias{expectBY}
\title{Expected B Y Product Calculation}
\usage{
expectBY(dgp, bases)
}
\arguments{
\item{dgp}{A list describing the data-generating process.}

\item{bases}{A list of basis functions.}
}
\value{
A vector representing the expected B Y product.
}
\description{
Computes the expected product of B and Y given a DGP and basis functions.
}
