% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeinfCode.R
\name{expectYConditionalOnDZ}
\alias{expectYConditionalOnDZ}
\title{Expected Outcome Given D and Z}
\usage{
expectYConditionalOnDZ(z, d, dgp)
}
\arguments{
\item{z}{The instrument value.}

\item{d}{The treatment indicator (0 or 1).}

\item{dgp}{A list describing the data-generating process.}
}
\value{
Expected outcome given D and Z.
}
\description{
Computes the expected outcome \deqn{E[Y \mid D, Z]}{E[Y | D, Z]} using MTR coefficients.
}
