% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeinfCode.R
\name{print.shapeinf}
\alias{print.shapeinf}
\title{Print Method for \code{shapeinf} Objects}
\usage{
\method{print}{shapeinf}(x, ci.digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{shapeinf}.}

\item{ci.digits}{Number of digits to display in confidence intervals.}

\item{...}{Additional arguments (not currently used).}
}
\description{
Print Method for \code{shapeinf} Objects
}
