% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeinf-package.R
\docType{package}
\name{shapeinf-package}
\alias{shapeinf}
\alias{shapeinf-package}
\title{shapeinf: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ian-xu-economics/shapeinf}
  \item Report bugs at \url{https://github.com/ian-xu-economics/shapeinf/issues}
}

}
\author{
\strong{Maintainer}: Ian Xu \email{ianxu@uchicago.edu} (\href{https://orcid.org/{{orcid}}}{ORCID}) [copyright holder]

}
\keyword{internal}
