% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeinfCode.R
\name{summary.shapeinf}
\alias{summary.shapeinf}
\title{Summary Method for \code{shapeinf} Objects}
\usage{
\method{summary}{shapeinf}(object, alpha = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{shapeinf}.}

\item{alpha}{Optional new significance level.}

\item{...}{Additional arguments (not currently used).}
}
\description{
Summary Method for \code{shapeinf} Objects
}
