% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeinf-package.R
\docType{package}
\name{shapeinf-package}
\alias{shapeinf-package}
\title{shapeinf: Shape Restricted Inference}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ian-xu-economics/shapeinf}
  \item Report bugs at \url{https://github.com/ian-xu-economics/shapeinf/issues}
}

}
\author{
\strong{Maintainer}: Ian Xu \email{ianxu@uchicago.edu}

Authors:
\itemize{
  \item Alexander Torgovitsky
}

}
\keyword{internal}
